requirements.txt
Flask
pywebview
opencv-python
numpy
mediapipe
joblib
pyautogui
scikit-learn